requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.4
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
attrs==25.3.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comm==0.2.2
comtypes==1.4.11
contourpy==1.3.1
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
dataclasses-json==0.6.7
debugpy==1.8.13
decorator==5.1.1
Distance==0.1.3
distro==1.9.0
executing==2.2.0
faiss-cpu==1.11.0
fonttools==4.55.7
frozenlist==1.6.0
g2p-en==2.1.0
greenlet==3.2.2
gTTS==2.5.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
inflect==7.5.0
ipyevents==2.0.2
ipykernel==6.29.5
ipython==8.31.0
ipywidgets==8.1.5
jedi==0.19.2
Jinja2==3.1.5
jiter==0.10.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
kiwisolver==1.4.8
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.61
langchain-openai==0.3.18
langchain-text-splitters==0.3.8
langsmith==0.3.43
lazy_loader==0.4
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.0
matplotlib-inline==0.1.7
mne==1.9.0
more-itertools==10.7.0
moviepy==2.2.1
multidict==6.4.4
mypy_extensions==1.1.0
nest-asyncio==1.6.0
nibabel==5.3.2
nltk==3.9.1
numpy==2.2.2
openai==1.82.0
opencv-python==4.11.0.86
orjson==3.10.18
packaging==24.2
pandas==2.2.3
parso==0.8.4
pillow==11.1.0
platformdirs==4.3.6
playsound==1.3.0
pooch==1.8.2
proglog==0.1.12
prompt_toolkit==3.0.50
propcache==0.3.1
psutil==7.0.0
pure_eval==0.2.3
pycairo==1.28.0
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.1
pyparsing==3.2.1
pyperclip==1.9.0
pypiwin32==223
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.16.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-vlc==3.0.21203
pyttsx3==2.98
pytz==2025.1
pyvista==0.44.2
pyvistaqt==0.11.1
pywin32==309
PyYAML==6.0.2
pyzmq==26.3.0
QtPy==2.4.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
scikit-learn==1.6.1
scipy==1.15.1
scooby==0.10.0
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
stack-data==0.6.3
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.9.0
tkintervideo==0.0.2
tkVideo==0.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typeguard==4.4.2
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
vtk==9.3.1
wcwidth==0.2.13
wheel==0.45.1
widgetsnbextension==4.0.13
yarl==1.20.0
zstandard==0.23.0